Fix weekly sensor.
Way you have it written, it's looking for the 31st of February. Which does not exist, and would be in the future if it does.

Instead, you want to get the timestamp for now and subtract 7 days from it.

Also, ideally this sensor would exclude time on home screen, netflix, and Hulu. But at that point it's probably an Influx query.